requirements.txt
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
backoff==2.2.1
bcrypt==5.0.0
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.44.1
colorama==0.4.6
coverage==7.13.0
Deprecated==1.2.14
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
googleapis-common-protos==1.61.0
greenlet==3.3.0
grpcio==1.59.3
h11==0.14.0
httpcore==1.0.8
httpx==0.25.1
idna==3.4
iniconfig==2.3.0
opentelemetry-api==1.12.0
opentelemetry-exporter-otlp==1.12.0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
opentelemetry-exporter-otlp-proto-http==1.12.0
opentelemetry-proto==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.1
pydantic_core==2.14.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
redis==5.0.1
requests==2.31.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
wrapt==1.16.0
